Inventory views created/modified.

+++ SQL/BMI.BI/BMI_BI_DW/Programmable Objects/discovery/Views/inventory_snapshot_m3v10_pol.sql
@@ -1,4 +1,4 @@
-﻿IF OBJECT_ID('[discovery].[inventory_snapshot_m3v10_pol]') IS NOT NULL
+IF OBJECT_ID('[discovery].[inventory_snapshot_m3v10_pol]') IS NOT NULL
 	DROP VIEW [discovery].[inventory_snapshot_m3v10_pol];
 
 GO
@@ -6,6 +6,11 @@ SET QUOTED_IDENTIFIER ON
 GO
 SET ANSI_NULLS ON
 GO
+
+
+
+
+
 --<CommentHeader>
 /**********************************************************************************************************************
 
@@ -36,30 +41,32 @@ Version	ChangeDate		Author	BugRef	Narrative
 create   view [discovery].[inventory_snapshot_m3v10_pol]
 as
 select
-	MITTRA.EtlCreatedOn						  as etl_created_on
-  , MITTRA.EtlSourceTable					  as etl_source_table
-  , max(MITTRA.MTTRDT)						  as transaction_date
-  , MITTRA.DataSourceKey					  as data_source_key
-  , ds.DataSourceName						  as data_source_name
-  , 'POL'									  as country_code
-  , MITTRA.MTWHLO							  as native_warehouse_number
-  , MITWHL.MWWHNM							  as native_warehouse_name
-  , MITTRA.MTITNO							  as native_product_key
-  , coalesce(prod.ProductName, MITMAS.MMITDS) as product_name
-  , prod.ProductCategoryBaseLevel			  as product_category_direct
-  , prod.ProductCategoryLevel1				  as product_category_level_1
-  , prod.ProductCategoryLevel2				  as product_category_level_2
-  , prod.ProductCategoryLevel3				  as product_category_level_3
-  , prod.ProductCategoryLevel4				  as product_category_level_4
-  , prod.ProductCategoryLevel5				  as product_category_level_5
-  , sum(MITTRA.MTTRQT)						  as native_stock_on_hand
-  , MITMAS.MMUNMS							  as native_unit_of_measure
-  , coalesce(uom.unit_of_measure_code, '')	  as standard_unit_of_measure
-  , 'PLN'									  as local_ccy
-  , ''										  as native_unit_value_local_ccy
-  , ''										  as stock_in_hand_value_local_ccy
-  , MITMAS.MMPUPR							  as native_unit_value_eur
-  , sum(MITTRA.MTTRQT * MITMAS.MMPUPR)		  as stock_in_hand_value_eur
+	MITTRA.EtlCreatedOn									  as etl_created_on
+  , MITTRA.EtlUpdatedOn									  as etl_updated_on
+  , MITTRA.EtlSourceTable								  as etl_source_table
+  , MITTRA.IsDeleted									  as is_deleted
+  , max(cast(cast(MITTRA.MTTRDT as char(8)) as datetime)) as snapshot_date
+  , MITTRA.DataSourceKey								  as data_source_key
+  , ds.DataSourceName									  as data_source_name
+  , 'POL'												  as country_code
+  , MITTRA.MTWHLO										  as native_warehouse_number
+  , MITWHL.MWWHNM										  as native_warehouse_name
+  , MITTRA.MTITNO										  as native_product_key
+  , coalesce(prod.ProductName, MITMAS.MMITDS)			  as product_name
+  , prod.ProductCategoryBaseLevel						  as product_category_direct
+  , prod.ProductCategoryLevel1							  as product_category_level_1
+  , prod.ProductCategoryLevel2							  as product_category_level_2
+  , prod.ProductCategoryLevel3							  as product_category_level_3
+  , prod.ProductCategoryLevel4							  as product_category_level_4
+  , prod.ProductCategoryLevel5							  as product_category_level_5
+  , sum(MITTRA.MTTRQT)									  as native_stock_on_hand
+  , MITMAS.MMUNMS										  as native_unit_of_measure
+  , coalesce(uom.unit_of_measure_code, '')				  as standard_unit_of_measure
+  , 'PLN'												  as local_ccy
+  , ''													  as native_unit_value_local_ccy
+  , ''													  as stock_in_hand_value_local_ccy
+  , MITMAS.MMPUPR										  as native_unit_value_eur
+  , sum(MITTRA.MTTRQT * MITMAS.MMPUPR)					  as stock_in_hand_value_eur
 from
 	psa.ics_stg_M3V10gen_POL_MITTRA		  as MITTRA ---- Material table
 inner join dbo.DataSource				  as ds
@@ -80,8 +87,8 @@ group by
 	MITTRA.MTWHLO
   , MITTRA.MTITNO
   , MITMAS.MMITDS
-  , MTWHSL
-  , MTBANO
+	--, MTWHSL
+	--, MTBANO
   , MITMAS.MMUNMS
   , MITWHL.MWWHNM
   , MITTRA.EtlCreatedOn
@@ -97,34 +104,38 @@ group by
   , ds.DataSourceName
   , coalesce(uom.unit_of_measure_code, '')
   , MITMAS.MMPUPR
+  , MITTRA.EtlUpdatedOn
+  , MITTRA.IsDeleted
 having
 	sum(MITTRA.MTTRQT) <> 0
 union
 select
-	MITTRA.EtlCreatedOn						  as etl_created_on
-  , MITTRA.EtlSourceTable					  as etl_source_table
-  , max(MITTRA.MTTRDT)						  as transaction_date
-  , MITTRA.DataSourceKey					  as data_source_key
-  , ds.DataSourceName						  as data_source_name
-  , 'POL'									  as country_code
-  , MITTRA.MTWHLO							  as native_warehouse_number
-  , MITWHL.MWWHNM							  as native_warehouse_name
-  , MITTRA.MTITNO							  as native_product_key
-  , coalesce(prod.ProductName, MITMAS.MMITDS) as product_name
-  , prod.ProductCategoryBaseLevel			  as product_category_direct
-  , prod.ProductCategoryLevel1				  as product_category_level_1
-  , prod.ProductCategoryLevel2				  as product_category_level_2
-  , prod.ProductCategoryLevel3				  as product_category_level_3
-  , prod.ProductCategoryLevel4				  as product_category_level_4
-  , prod.ProductCategoryLevel5				  as product_category_level_5
-  , sum(MITTRA.MTTRQT)						  as native_stock_on_hand
-  , MITMAS.MMUNMS							  as native_unit_of_measure
-  , coalesce(uom.unit_of_measure_code, '')	  as standard_unit_of_measure
-  , 'PLN'									  as local_ccy
-  , ''										  as native_unit_value_local_ccy
-  , ''										  as stock_in_hand_value_local_ccy
-  , MITMAS.MMPUPR							  as native_unit_value_eur
-  , sum(MITTRA.MTTRQT * MITMAS.MMPUPR)		  as stock_in_hand_value_eur
+	MITTRA.EtlCreatedOn									  as etl_created_on
+  , MITTRA.EtlUpdatedOn									  as etl_updated_on
+  , MITTRA.EtlSourceTable								  as etl_source_table
+  , MITTRA.IsDeleted									  as is_deletede
+  , max(cast(cast(MITTRA.MTTRDT as char(8)) as datetime)) as transaction_date
+  , MITTRA.DataSourceKey								  as data_source_key
+  , ds.DataSourceName									  as data_source_name
+  , 'POL'												  as country_code
+  , MITTRA.MTWHLO										  as native_warehouse_number
+  , MITWHL.MWWHNM										  as native_warehouse_name
+  , MITTRA.MTITNO										  as native_product_key
+  , coalesce(prod.ProductName, MITMAS.MMITDS)			  as product_name
+  , prod.ProductCategoryBaseLevel						  as product_category_direct
+  , prod.ProductCategoryLevel1							  as product_category_level_1
+  , prod.ProductCategoryLevel2							  as product_category_level_2
+  , prod.ProductCategoryLevel3							  as product_category_level_3
+  , prod.ProductCategoryLevel4							  as product_category_level_4
+  , prod.ProductCategoryLevel5							  as product_category_level_5
+  , sum(MITTRA.MTTRQT)									  as native_stock_on_hand
+  , MITMAS.MMUNMS										  as native_unit_of_measure
+  , coalesce(uom.unit_of_measure_code, '')				  as standard_unit_of_measure
+  , 'PLN'												  as local_ccy
+  , ''													  as native_unit_value_local_ccy
+  , ''													  as stock_in_hand_value_local_ccy
+  , MITMAS.MMPUPR										  as native_unit_value_eur
+  , sum(MITTRA.MTTRQT * MITMAS.MMPUPR)					  as stock_in_hand_value_eur
 from
 	psa.ics_stg_M3V10ved_POL_MITTRA		  as MITTRA ---- Material table
 inner join dbo.DataSource				  as ds
@@ -145,8 +156,8 @@ group by
 	MITTRA.MTWHLO
   , MITTRA.MTITNO
   , MITMAS.MMITDS
-  , MTWHSL
-  , MTBANO
+	--, MTWHSL
+	--, MTBANO
   , MITMAS.MMUNMS
   , MITWHL.MWWHNM
   , MITTRA.EtlCreatedOn
@@ -162,34 +173,38 @@ group by
   , ds.DataSourceName
   , coalesce(uom.unit_of_measure_code, '')
   , MITMAS.MMPUPR
+  , MITTRA.EtlUpdatedOn
+  , MITTRA.IsDeleted
 having
 	sum(MITTRA.MTTRQT) <> 0
 union
 select
-	MITTRA.EtlCreatedOn						  as etl_created_on
-  , MITTRA.EtlSourceTable					  as etl_source_table
-  , max(MITTRA.MTTRDT)						  as transaction_date
-  , MITTRA.DataSourceKey					  as data_source_key
-  , ds.DataSourceName						  as data_source_name
-  , 'POL'									  as country_code
-  , MITTRA.MTWHLO							  as native_warehouse_number
-  , MITWHL.MWWHNM							  as native_warehouse_name
-  , MITTRA.MTITNO							  as native_product_key
-  , coalesce(prod.ProductName, MITMAS.MMITDS) as product_name
-  , prod.ProductCategoryBaseLevel			  as product_category_direct
-  , prod.ProductCategoryLevel1				  as product_category_level_1
-  , prod.ProductCategoryLevel2				  as product_category_level_2
-  , prod.ProductCategoryLevel3				  as product_category_level_3
-  , prod.ProductCategoryLevel4				  as product_category_level_4
-  , prod.ProductCategoryLevel5				  as product_category_level_5
-  , sum(MITTRA.MTTRQT)						  as native_stock_on_hand
-  , MITMAS.MMUNMS							  as native_unit_of_measure
-  , coalesce(uom.unit_of_measure_code, '')	  as standard_unit_of_measure
-  , 'PLN'									  as local_ccy
-  , ''										  as native_unit_value_local_ccy
-  , ''										  as stock_in_hand_value_local_ccy
-  , MITMAS.MMPUPR							  as native_unit_value_eur
-  , sum(MITTRA.MTTRQT * MITMAS.MMPUPR)		  as stock_in_hand_value_eur
+	MITTRA.EtlCreatedOn									  as etl_created_on
+  , MITTRA.EtlUpdatedOn									  as etl_updated_on
+  , MITTRA.EtlSourceTable								  as etl_source_table
+  , MITTRA.IsDeleted									  as is_deleted
+  , max(cast(cast(MITTRA.MTTRDT as char(8)) as datetime)) as transaction_date
+  , MITTRA.DataSourceKey								  as data_source_key
+  , ds.DataSourceName									  as data_source_name
+  , 'POL'												  as country_code
+  , MITTRA.MTWHLO										  as native_warehouse_number
+  , MITWHL.MWWHNM										  as native_warehouse_name
+  , MITTRA.MTITNO										  as native_product_key
+  , coalesce(prod.ProductName, MITMAS.MMITDS)			  as product_name
+  , prod.ProductCategoryBaseLevel						  as product_category_direct
+  , prod.ProductCategoryLevel1							  as product_category_level_1
+  , prod.ProductCategoryLevel2							  as product_category_level_2
+  , prod.ProductCategoryLevel3							  as product_category_level_3
+  , prod.ProductCategoryLevel4							  as product_category_level_4
+  , prod.ProductCategoryLevel5							  as product_category_level_5
+  , sum(MITTRA.MTTRQT)									  as native_stock_on_hand
+  , MITMAS.MMUNMS										  as native_unit_of_measure
+  , coalesce(uom.unit_of_measure_code, '')				  as standard_unit_of_measure
+  , 'PLN'												  as local_ccy
+  , ''													  as native_unit_value_local_ccy
+  , ''													  as stock_in_hand_value_local_ccy
+  , MITMAS.MMPUPR										  as native_unit_value_eur
+  , sum(MITTRA.MTTRQT * MITMAS.MMPUPR)					  as stock_in_hand_value_eur
 from
 	psa.ics_stg_M3V10vil_POL_MITTRA		  as MITTRA ---- Material table
 inner join dbo.DataSource				  as ds
@@ -210,8 +225,8 @@ group by
 	MITTRA.MTWHLO
   , MITTRA.MTITNO
   , MITMAS.MMITDS
-  , MTWHSL
-  , MTBANO
+	--, MTWHSL
+	--, MTBANO
   , MITMAS.MMUNMS
   , MITWHL.MWWHNM
   , MITTRA.EtlCreatedOn
@@ -227,6 +242,8 @@ group by
   , ds.DataSourceName
   , coalesce(uom.unit_of_measure_code, '')
   , MITMAS.MMPUPR
+  , MITTRA.EtlUpdatedOn
+  , MITTRA.IsDeleted
 having
 	sum(MITTRA.MTTRQT) <> 0 ;
 GO
